+++ app/middlewares/users.js
@@ -1,6 +1,6 @@
 const jwt = require('jwt-simple');
-
 const { checkSchema } = require('express-validator');
+
 const checkValidations = require('../helpers/checkValidation');
 
 exports.session = (request, response, next) => {
@@ -75,33 +75,6 @@ exports.create = (request, response, next) => {
         // always return false
         options: () => false
       }
-    }
-  });
-
-  checkValidations(request, response, next, validations);
-};
-
-exports.list = (request, response, next) => {
-  const validations = checkSchema({
-    page: {
-      in: 'query',
-      optional: true,
-      isInt: {
-        options: {
-          gt: 0
-        }
-      },
-      errorMessage: 'page must be a number greater than zero (0)'
-    },
-    limit: {
-      in: 'query',
-      optional: true,
-      isInt: {
-        options: {
-          gt: 0
-        }
-      },
-      errorMessage: 'limit must be a number greater than zero (0)'
     }
   });
